Move project.json to /src instead of /public
