--- public/data/projets.json
@@ -1,31 +1,0 @@
- [
-    {
-        "id" :"1",
-        "image":"nounouproject.png",
-        "name": "Nounou Project",
-        "description": "Site pour faciliter la communication entre une assistante maternelle et les parents",
-        "techno":"React.js / node.js / Express / MySQL  ",
-        "repo":"https://github.com/Angelina2504/nounouproject"
-    
-    },
-    {
-        "id" :"2",
-        "image":"homeNoMade.png",
-        "name": "No-made",
-        "description": "Proposer un site de connection à la nature en intérieur et en exterieur.",
-        "techno":"React.js / module.CSS",
-        "repo":"https://github.com/Angelina2504/no_made",
-        "site":"https://no-made.vercel.app/"
-    
-    },
-    {
-        "id" :"3",
-        "image":"homeLegoMuscu.png",
-        "name": "Lego Muscu",
-        "description": "Trouver son coach sportif en fonction des résultats du questionnaire",
-        "techno":"HTML / CSS / JS",
-        "repo": "https://github.com/WildCodeSchool-2024-02/JS-RemoteFR-P1-Lego-Muscu.git",
-        "site":"https://wildcodeschool-2024-02.github.io/JS-RemoteFR-P1-Lego-Muscu/"
-        
-    }
-]
